+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 All notable changes to the "pastum" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+<!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
+
+## [0.0.2] - 2024-10-26
+
+- Addded distinction between Integer/Float values
+- Missing, NA and None values inserted by default if the value empty
 
 ## [0.0.1] - 2024-10-24